Clean up shaders

+++ assets/shaders/stage1.wgsl
@@ -1,5 +1,5 @@
-#!import cells
 #!import hex
+#!import cells
 
 @group(0) @binding(0) var <storage, read_write> cells: Cells;
 
@@ -7,7 +7,7 @@
 @workgroup_size(8, 8, 1)
 fn main(@builtin(global_invocation_id) id: vec3u) {
   let axial: vec2i = vec2i(id.xy) - i32(cells.radius);
-  if(!isInBounds(axial, i32(cells.radius))){
+  if(!isInBounds(axial, cells.radius)){
     return;
   }
