+++ assets/shaders/stage2.wgsl
@@ -1,5 +1,5 @@
-#!import cells
 #!import hex
+#!import cells
 
 struct Settings {
   alpha: f32,
@@ -8,19 +8,21 @@ struct Settings {
 }
 
 struct RenderCellsSettings {
+  // stores bits of f32 maxValue
   maxValue: atomic<u32>,
 }
 
 @group(0) @binding(0) var <uniform> settings: Settings;
 @group(0) @binding(1) var <storage, read_write> cells: Cells;
 @group(0) @binding(2) var <storage, read_write> renderSettings: RenderCellsSettings;
+// as a bool
 @group(0) @binding(3) var <storage, read_write> finished: u32;
 
 @compute
 @workgroup_size(8, 8, 1)
 fn main(@builtin(global_invocation_id) id: vec3u) {
   let axial: vec2i = vec2i(id.xy) - i32(cells.radius);
-  if(!isInBounds(axial, i32(cells.radius)) || finished == 1){
+  if(!isInBounds(axial, cells.radius) || finished == 1){
     return;
   }
 
@@ -35,16 +37,27 @@ fn main(@builtin(global_invocation_id) id: vec3u) {
     let neighbourPosition: vec2i = cellNeighbours[i];
     let neighbourIndex: u32 = getCellIndex(neighbourPosition);
     let neighbour: Cell = cells.cells[neighbourIndex];
-    let neighbourValue: f32 = select(settings.beta, getValue(&cells.cells[neighbourIndex]), isInBounds(neighbourPosition, i32(cells.radius)));
+    let neighbourValue: f32 = 
+      select(
+        settings.beta,
+        getValue(&cells.cells[neighbourIndex]),
+        isInBounds(neighbourPosition, cells.radius),
+      );
     
     diffusion += select(neighbourValue / 12.0, 0.0, neighbour.receptive == 1);
   }
 
   let newValue: f32 = value + diffusion;
-  let maxValue: f32 = max(bitcast<f32>(atomicLoad(&renderSettings.maxValue)), newValue);
+  let maxValue: f32 = max(
+    bitcast<f32>(atomicLoad(&renderSettings.maxValue)),
+    newValue
+  );
   atomicStore(&renderSettings.maxValue, bitcast<u32>(maxValue));
 
-  finished = max(finished, select(0u, 1u, axialRadius(axial) == cells.radius && newValue >= 1.0));
+  finished = max(
+    finished,
+    select(0u, 1u, axialRadius(axial) == cells.radius && newValue >= 1.0)
+  );
 
   setValue(&cells.cells[index], newValue, cells.useValue);
   setDiffusion(&cells.cells[index], diffusion, cells.useValue);
